+++ src/secantus/__init__.py
@@ -1,5 +1,5 @@
 from secantus.server import SecantusDBServer
 
-__version__ = "0.5.4b127"
+__version__ = "0.5.4b128"
 
 __all__ = ["SecantusDBServer", "__version__"]